Works with PDBs now.
Check in before optimization and beuatification

+++ Slider/Interfaces/IGenerator.cs
@@ -6,6 +6,6 @@ namespace Slider.Interfaces
 {
     public interface IGenerator
     {
-        List<int> Generate(int gridSize);
+        List<byte> Generate(int gridSize);
     }
 }
